requirements.txt
# Core
youtube-transcript-api>=0.6.1
sentence-transformers>=2.2.2
keybert>=0.7.0
torch>=2.0.0
nltk>=3.8.1
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.0.0
anthropic>=0.25.0

# Audio processing
yt-dlp>=2024.3.10
openai-whisper>=20231117

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Code Quality
black>=24.0.0
ruff>=0.4.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0